extract SshExecutor into own file
